runtime: clean-up types at Stanza boundary
Introduce `stz_*` types to explicitly control size/signedness of types
that may be passed to/from Stanza. This is important because Stanza
types and C types have similar names, but not necessarily the same
representation.

For example, Stanza `long` is a 64-bit signed integer. However, its
size/signedness in C is platform specific. Indeed, on Windows it is a
32-bit signed integer, and on Linux it is a 64-bit signed integer. By
using the consistently-defined type `stz_long` we know that our
representation in C will match the representation of `long` in Stanza.