runtime: windows: better `exit_with_error()`.
On Windows, `errno` is not typically set when Windows API functions
fail. Hence `strerror(errno)` yields the ever-so-helpful error message
of "No error".

Instead, use `FormatMessage` and `GetLastError` to actually print out
the error that occurred. These are Windows API functions that will
format and retrieve the previous error. This way we get a useful error
message when things fail.