runtime: windows: shrinking memory regions
Correctly implement shrinking memory regions by decommitting pages that
are no longer in use.

Previously, we were not actually decommitting when shrinking, because
that appeared to cause more than we wanted to be decommitted. The issue
was that we were not rounding to a page boundary. Hence for some
shrinking operations that did not shrink more than a page (or cross a
page boundary) we would inadvertently decommit a page that still had
live data. Rounding up to the nearest page boundary and checking that
they are not the same (that there are pages to decommit) fixes this
issue.