core: windows: implement line prompter for REPL
Since linenoise is not available on Windows, we can't use it. Instead,
just implement a simple line prompter that prints the prompt, takes
input and then gives it back to the REPL. This doesn't get us any fancy
behavior like command history, but it should get the job done for now.