runtime: windows: don't call `CloseHandle()` twice on the same handle.
In order to keep good handle hygiene, we make sure to close the
inherited handles that we pass into the process (`stdin_read`,
`stdout_write` and `stderr_write`).

However, in the event that we redirect both output streams of a process
to the same stream (i.e. `STDOUT/STDERR` to `PROCESS-OUT` or
`STDOUT/STDERR` to `PROCESS-ERR`), we need to take care not to
inadvertently close the same file handle twice. Since `stdout_write` and
`stderr_write` point to the same handle, calling `CloseHandle()` on both
causes us to close the same handle twice. This causes an exception in
the Windows runtime, (sometimes) leading to a segfault upstream.

In order to prevent this scenario, we employ a trick similar to what we
used to set up the file handles. We keep track of three pointers that
correspond to a potential input/output/error handle, and then assign
these to the corresponding redirected handle. We then close
the pointers to valid handles, knowing that if we redirected both
streams to the same handle, the second redirection would overwrite the
handle value for the first, hence no double-close will occur.