runtime: windows: duplicate pipe handles we pass into the child process
When we create pipes for redirecting a child process's
input/output/error streams, we have to take special care to duplicate
any pipe ends that may be shared by other streams. This is to prevent
two things:

1. Two streams receiving the same handle to a pipe (technically invalid)
2. Closing the same pipe handle twice when attempting to close the ends
   of the stream that we passed into the child (causes an exception in
   the Windows runtime)

For example, with the original code, if we were to redirect both STDOUT
and STDERR to PROCESS-OUT, we would inadvertently end up with both
STDOUT and STDERR in the child process having the same handle to the end
of the PROCESS-OUT pipe, and then call `CloseHandle()` on those handles
in the parent process, only to inadvertently close the same handle
twice. Duplicating the handles ensures that both streams have a handle
to the same pipe end, but not the same handle, resolving both of the
above issues.