all: introduce the `include` dir and `stanza.h`
Add an `include` dir that contains Stanza-specific headers which should
be used when writing C code that interfaces with Stanza. This should be
the mechanism by which C code accesses the mechanisms of the runtime and
also allows re-use of types and constructs that the runtime uses to
match the exact representation of Stanza types and constructs (e.g.
`stz_long`)

Additionally, add a `stanza.h` header which is the amalgamation of all
the specific bits that are exposed in each Stanza header.  Currently
this only comprises `stanza/types.h` which contains typedefs for Stanza
types and some basic convenience macros for dealing with strings and
integer literals.

Since the runtime uses these headers, and we compile the runtime into
every Stanza exectuable, we now also add a `-I include' flag to the
compile line. This has the added benefit of automatically making the
Stanza headers available to any ccfiles that may be passed in on the
same command line. This way users have two separate, but equally
functional, methods of compiling C code with Stanza. Either the user
compiles their C code separately while passing `-I <StanzaIncludeDir>`,
or passes it as a `-ccfile` where it will be added onto the compile line
automatically.