stitcher: generate `EXPORTS` table on Windows
In order for externally-visible symbols to be imported by the dynamic
loader, they must be specifically placed into an `EXPORTS` table by the
assembler. This is done via the `drectve` section which contains a list
of null-terminated strings of the form "-export:\"<symbol>\"", where
`<symbol>` denotes the name of the symbol to-be-exported.

On Windows, we generate an entry in the `EXPORTS` table for each `extern
defn` symbol in each package, after all the other tables have been
generated.